fix: prevent endless loops
If the service worker code doesn't intercept a request  properly
we can end up redirecting back to the current page endlessly.

Instead use local storage to count the number of times we have
redirected to the current page and show an error message if it's
too many.
